refactor(mobile): split write through
- extract command and policy modules
- add write-through source-contract test
